+++ resources/[dg]/dg-vehicles/typescript/src/client/modules/status/service.status.ts
@@ -1,16 +1,12 @@
-import { Events, RPC, Sounds, UI, Util } from '@dgx/client';
+import { Events, RPC, Sounds, Util } from '@dgx/client';
 
 import { getCurrentVehicle, isDriver } from '../../helpers/vehicle';
-import { getCurrentWorkingShop } from '../mechanic/service.mechanic';
-import { getPerformanceUpgrades } from '../upgrades/service.upgrades';
 
 import {
   MINIMUM_DAMAGE_FOR_GUARANTEED_STALL,
   MINIMUM_DAMAGE_FOR_STALL,
   degradationValues,
   handlingOverrideFunctions,
-  partNames,
-  serviceConditions,
 } from './constant.status';
 import { getVehicleFuel, overrideSetFuel } from 'modules/fuel/service.fuel';
 import { getVehicleVin } from 'modules/identification/service.identification';
@@ -18,7 +14,7 @@ import { tryEjectAfterCrash } from 'modules/seatbelts/service.seatbelts';
 import { setEngineState } from 'services/engine';
 import { hasVehicleKeys } from 'modules/keys/cache.keys';
 
-const vehicleService: {
+let vehicleService: {
   vehicle: number;
   numWheels: number;
   vin: string;
@@ -26,22 +22,11 @@ const vehicleService: {
     status: NodeJS.Timer;
     apply: NodeJS.Timer;
     sync: NodeJS.Timer;
-  } | null;
-  info: Service.Status | null;
+  };
+  info: Service.Status;
   state: Record<'engine' | 'body', number>;
   originalHandling: Record<string, number>;
-} = {
-  vehicle: 0,
-  numWheels: 0,
-  vin: '',
-  threads: null,
-  info: null,
-  state: {
-    engine: 1000,
-    body: 1000,
-  },
-  originalHandling: {},
-};
+} | null = null;
 
 const multipliers = {
   brake: 1,
@@ -55,30 +40,40 @@ const clearOverrideThreads = () => {
   overrideThreads = [];
 };
 
-// region Service satuts
-const saveHandlingValues = (veh: number) => {
+// region Service status
+const getOriginalHandlingValues = (vehicle: number) => {
+  // if already in statebag, use those. else get from native and save in statebag
+  const entState = Entity(vehicle).state;
+  const handlingValuesFromState = entState.handlingValues;
+  if (handlingValuesFromState) return handlingValuesFromState;
+
+  const originalHandling: Record<string, number> = {};
+
   for (const part in degradationValues) {
     for (const value of degradationValues[part as keyof Service.Status]) {
       const isOverride = !!handlingOverrideFunctions[value.name];
-      const original = isOverride ? 1 : GetVehicleHandlingFloat(veh, 'CHandlingData', value.name);
-      vehicleService.originalHandling[value.name] = original;
+      const original = isOverride ? 1 : GetVehicleHandlingFloat(vehicle, 'CHandlingData', value.name);
+      originalHandling[value.name] = original;
     }
   }
+
+  entState.set('handlingValues', originalHandling, false);
+
+  return originalHandling;
 };
 
-const resetHandlingValues = (veh: number) => {
-  for (const [hKey, hValue] of Object.entries(vehicleService.originalHandling)) {
-    if (handlingOverrideFunctions[hKey]) return;
-    SetVehicleHandlingFloat(veh, 'CHandlingData', hKey, hValue);
+const applyHandlingValues = (vehicle: number, handlingValues: Record<string, number>) => {
+  for (const [key, value] of Object.entries(handlingValues)) {
+    if (handlingOverrideFunctions[key]) return;
+    SetVehicleHandlingFloat(vehicle, 'CHandlingData', key, Number(value));
   }
   clearOverrideThreads();
-  vehicleService.originalHandling = {};
 };
 
-const calculateDegrationSteps = () => {
+const calculateDegrationSteps = (handlingValues: Record<string, number>) => {
   for (const part in degradationValues) {
     for (const value of degradationValues[part as keyof Service.Status]) {
-      const original = vehicleService.originalHandling[value.name];
+      const original = handlingValues[value.name];
       value.bottom = original * value.percent;
       value.step = (original - value.bottom) / 1000;
     }
@@ -86,7 +81,7 @@ const calculateDegrationSteps = () => {
 };
 
 const setVehicleDegradation = (veh: number) => {
-  if (!vehicleService.info) return;
+  if (!vehicleService) return;
 
   clearOverrideThreads();
 
@@ -109,7 +104,7 @@ const setVehicleDegradation = (veh: number) => {
 
 // Thread gets started whenever player entered driver seat of a vehicle
 export const startStatusThread = async (vehicle: number) => {
-  if (vehicleService.threads !== null) {
+  if (vehicleService !== null) {
     console.error('Status threads already running, canceling');
     return;
   }
@@ -118,48 +113,39 @@ export const startStatusThread = async (vehicle: number) => {
   const vin = await getVehicleVin(vehicle);
   if (!netId || !vin) return;
 
-  vehicleService.vehicle = vehicle;
-  vehicleService.numWheels = GetVehicleNumberOfWheels(vehicle);
-  vehicleService.vin = vin;
-  vehicleService.info = await RPC.execute('vehicles:service:getStatus', netId);
-  vehicleService.state.engine = GetVehicleEngineHealth(vehicle);
-  vehicleService.state.body = GetVehicleBodyHealth(vehicle);
+  const info = await RPC.execute('vehicles:service:getStatus', netId);
+  const state = {
+    engine: GetVehicleEngineHealth(vehicle),
+    body: GetVehicleBodyHealth(vehicle),
+  };
+
+  const originalHandling = getOriginalHandlingValues(vehicle);
 
   // Calc handling values
-  saveHandlingValues(vehicle);
-  calculateDegrationSteps();
-
+  calculateDegrationSteps(originalHandling);
   setVehicleDegradation(vehicle);
-  vehicleService.threads = {
+
+  const threads = {
     // Statusthread calculates the vehicle status every 100ms
     status: setInterval(() => {
-      const currentVehicle = getCurrentVehicle();
-      if (
-        currentVehicle === null ||
-        !DoesEntityExist(currentVehicle) ||
-        GetPedInVehicleSeat(currentVehicle, -1) !== PlayerPedId()
-      ) {
-        cleanStatusThread();
-        return;
-      }
-
-      if (!vehicleService.info) return;
+      if (vehicle === null || !DoesEntityExist(vehicle) || GetPedInVehicleSeat(vehicle, -1) !== PlayerPedId()) return;
+      if (!vehicleService) return;
 
       const brakePressure: number[] = [];
       const suspCompress: number[] = [];
       for (let i = 0; i < vehicleService.numWheels; i++) {
-        brakePressure.push(GetVehicleWheelBrakePressure(currentVehicle, i));
-        suspCompress.push(GetVehicleWheelSuspensionCompression(currentVehicle, i));
+        brakePressure.push(GetVehicleWheelBrakePressure(vehicle, i));
+        suspCompress.push(GetVehicleWheelSuspensionCompression(vehicle, i));
       }
 
       // Brakes
       // Brake natives give wrong values when engine is off
-      if (GetIsVehicleEngineRunning(currentVehicle)) {
+      if (GetIsVehicleEngineRunning(vehicle)) {
         const avgBrakePressure = Util.average(brakePressure);
         if (avgBrakePressure <= 0) {
           multipliers.brake = 1;
         } else {
-          const brakeDecrease = (avgBrakePressure * multipliers.brake) / 20;
+          const brakeDecrease = (avgBrakePressure * multipliers.brake) / 40;
           vehicleService.info.brakes = Math.max(vehicleService.info.brakes - brakeDecrease, 0);
           multipliers.brake += 0.1;
         }
@@ -168,137 +154,71 @@ export const startStatusThread = async (vehicle: number) => {
       // Suspension
       const avgSuspCompress = Util.average(suspCompress);
       if (avgSuspCompress > 0.15) {
-        const suspDecrease = avgSuspCompress / 10;
+        const suspDecrease = avgSuspCompress / 30;
         vehicleService.info.suspension = Math.max(vehicleService.info.suspension - suspDecrease, 0);
       }
 
       // Axle & engine
-      const newBody = GetVehicleBodyHealth(currentVehicle);
+      const newBody = GetVehicleBodyHealth(vehicle);
       const bodyDelta = vehicleService.state.body - newBody;
       vehicleService.state.body = newBody;
 
-      const newEngine = GetVehicleEngineHealth(currentVehicle);
+      const newEngine = GetVehicleEngineHealth(vehicle);
       const engineDelta = vehicleService.state.engine - newEngine;
       vehicleService.state.engine = newEngine;
 
       if (bodyDelta > 0) {
-        const axleDecrease = bodyDelta / 6;
+        const axleDecrease = bodyDelta / 25;
         vehicleService.info.axle = Math.max(vehicleService.info.axle - axleDecrease, 0);
       }
       if (engineDelta > 0) {
-        const engineDecrease = engineDelta / 6;
+        const engineDecrease = engineDelta / 15;
         vehicleService.info.engine = Math.max(vehicleService.info.engine - engineDecrease, 0);
       }
     }, 100),
-    // Apply thread applies calculated service values every 5 seconds
+    // Every 5 seconds we apply calculated service degradation
     apply: setInterval(() => {
+      if (!vehicleService) return;
       setVehicleDegradation(vehicle);
     }, 5000),
     // Every minute we sync calculated service values
     sync: setInterval(() => {
-      if (!vehicleService.info) return;
-      Events.emitNet('vehicles:service:updateStatus', vehicleService.vin, vehicleService.info);
+      if (!vehicleService) return;
+      Events.emitNet('vehicles:service:saveStatus', vehicleService.vin, vehicleService.info);
     }, 60000),
   };
+
+  // save data for later
+  vehicleService = {
+    vehicle,
+    numWheels: GetVehicleNumberOfWheels(vehicle),
+    vin,
+    threads,
+    info,
+    state,
+    originalHandling,
+  };
+
+  // init degradation
+  setVehicleDegradation(vehicle);
 };
 
-const cleanStatusThread = () => {
-  Events.emitNet('vehicles:service:updateStatus', vehicleService.vin, vehicleService.info);
+export const cleanStatusThread = () => {
+  if (!vehicleService) return;
+
+  Events.emitNet('vehicles:service:saveStatus', vehicleService.vin, vehicleService.info);
 
   if (DoesEntityExist(vehicleService.vehicle)) {
-    resetHandlingValues(vehicleService.vehicle);
+    applyHandlingValues(vehicleService.vehicle, vehicleService.originalHandling);
   }
 
   if (vehicleService.threads) {
     for (const thread of Object.values(vehicleService.threads)) {
       clearInterval(thread);
     }
-    vehicleService.threads = null;
   }
 
-  vehicleService.vehicle = 0;
-  vehicleService.numWheels = 0;
-  vehicleService.vin = '';
-  vehicleService.info = null;
-};
-
-export const openServiceStatusOverview = async (veh: number) => {
-  if (!DoesEntityExist(veh)) return;
-  const serviceInfo = await RPC.execute<Service.Status>(
-    'vehicles:service:getStatus',
-    NetworkGetNetworkIdFromEntity(veh)
-  );
-  if (!serviceInfo) return;
-
-  const partMenu: ContextMenu.Entry[] = [];
-  let part: keyof Service.Status;
-  const plyShop = getCurrentWorkingShop();
-  for (part of Object.keys(serviceInfo) as (keyof Service.Status)[]) {
-    const partState = serviceInfo[part];
-    const partPerc = partState / 10;
-    const serviceCondition = serviceConditions.find(sc => sc.percentage <= partPerc);
-    partMenu.push({
-      title: `${partNames[part]}`,
-      description: plyShop
-        ? `${partPerc.toFixed(2)}% | ${Math.ceil((100 - partPerc) / 10) || 'no'} parts needed`
-        : `${serviceCondition?.label} condition`,
-    });
-  }
-
-  // Class info
-  const info = await RPC.execute<Upgrades.Performance & { class: CarClass; name: string }>(
-    'vehicles:service:getVehicleInfo',
-    NetworkGetNetworkIdFromEntity(veh)
-  );
-  const upgrades = getPerformanceUpgrades(veh);
-  const infoMenu: ContextMenu.Entry[] = info
-    ? [
-        {
-          title: info.name,
-          description: `Class: ${info.class}`,
-        },
-      ]
-    : [];
-
-  // Only show perf tunes to mechanic
-  if (plyShop) {
-    infoMenu.push(
-      ...[
-        {
-          title: 'Brakes',
-          description: (upgrades?.brakes ?? -1) === -1 ? 'Basis' : `Level ${upgrades!.brakes + 1}`,
-        },
-        {
-          title: 'Engine',
-          description: (upgrades?.engine ?? -1) === -1 ? 'Basis' : `Level ${upgrades!.engine + 1}`,
-        },
-        {
-          title: 'Transmission',
-          description: (upgrades?.transmission ?? -1) === -1 ? 'Basis' : `Level ${upgrades!.transmission + 1}`,
-        },
-        {
-          title: 'Turbo',
-          description: upgrades?.turbo ? `Geinstalleerd` : 'Niet geinstalleerd',
-        },
-        {
-          title: 'Suspension',
-          description: (upgrades?.suspension ?? -1) === -1 ? 'Basis' : `Level ${upgrades!.suspension + 1}`,
-        },
-      ]
-    );
-  }
-
-  const menu: ContextMenu.Entry[] = [
-    {
-      title: 'Voertuig info',
-      submenu: infoMenu,
-    },
-    {
-      title: 'Voertuig status',
-      submenu: partMenu,
-    },
-  ];
-  UI.openApplication('contextmenu', menu);
+  vehicleService = null;
 };
 // endregion