+++ resources/[dg]/dg-vehicles/typescript/src/types/service.d.ts
@@ -15,5 +15,6 @@ declare namespace Service {
 
   interface Config {
     degradationValues: DegradationConfig;
+    repairPercentagePerPart: number;
   }
 }